refactor: move findClosestAncestor logic to ApproachLevel class
